+++ lib/models/dog_model.dart
@@ -10,7 +10,7 @@ class DogModel extends PetModel {
     required String imageUrl,
     required DateTime birthDay,
     required GENDER_TYPE genderType,
-    required int weight,
+    required double weight,
     bool? isNeuter,
     bool? isPregnancy,
   }) : super(
